Move ContextPropagationDebug to internal package
And use it in the Instrumenter API
